Variable Explorer: Move set_value and get_value methods out of NamespaceBrowser
